Update README.md
inspiration taken from the stsquad fork to include both methods of installation. this is also in preparation for a separate gentoo-snappy fork that i'll use to make pull requests to the other forks, in order to help encourage developmental activity. although it appears the main fork is dead, perhaps some new life can be breathed into the other forks that show more recent activity.